#132 공개 일정 가져오기 메소드 이동 및 PlanService에서 제거
